Activity 3 Textbook Example 3 Navbar Date and Time
Create template index.html that matches example 3 in the textbook, which
includes a navbar, a callout of "Hello [my name]!" and a timestamp
reporter. This commit utilizes flask_bootstrap for styling and
flask_moment for reporting the time.

Modify hello.py to use pre-defined html templates.

Document screenshot of example 3's end result in README file.