Activity 4-1 Implement NameForm with flash warning message
Implement examples 4-1 to 4-7 in the textbook, which produces a form
that includes a required name entry and updates the header accordingly.
It also produces a warning when the user changes their input name.

Created 2 new files for this instead of modifying the existing hello.py
and index.html due to the 2 activities different nature and for
reference clarity.