Activity 4-2 Add EmailField to FlaskForm
Modify the original FlaskForm to contain an email field, which detects
whether the entry is a valid email address (i.e. includes '@') and also
whehter the email entered is a uoft email by detecting the string
"utoronto".

The website will generate different outcomes based on the email entry:

- Initial startup: No message shown
- UofT email: Display message "Your UofT email is {email}"
- Non-valid email: Show warning message
- Non-UofT email: Display message "Please use your UofT email."

Include 4 screenshots of the outcome of the 4 testing scenarios above in
the README file.